Adjustments > Levels: UI overhaul
This should *finally* bring the Levels dialog into line with other photo
editors.  Apologies that it took me so long to fix this.

Still to-do on this dialog:

- "Auto levels" button
- Color selectors for input shadow/highlight tones
- Toggle for logarithmic vs regular histogram
- Possibly separate channel handling.  I have already included code for
generating and rendering individual RGB histograms.  However, some deep
changes would be required to the Levels mapping itself, so I'm holding
off until I've gotten a better handle on how extensive the code changes
would need to be.