pdImage: convert PNG background color to dictionary storage
I forgot to mention it in the previous commit, but this is another great
advantage of using a dictionary for rare data entries - we only allocate
storage space for them if we absolutely need to.

Thanks to this, if say a JPEG is loaded, we don't waste space on entries
like this one (PNG background color).